faucet implemented

+++ client/Assets/Scripts/StructureManager.cs
@@ -24,7 +24,7 @@ public class StructureManager : MonoBehaviour
             return;
         if (placementManager.CheckIfPositionIsFree(position) == false)
         {
-            ContractManager.Instance.DeleteItem(position);
+            if (fromUser) { ContractManager.Instance.DeleteItem(position); }
             ContractManager.Instance.updateEditedMap(position, CellType.Empty);
             placementManager.RemoveItem(position);
             AudioPlayer.instance.PlayPlacementSound();
@@ -35,7 +35,7 @@ public class StructureManager : MonoBehaviour
     {
         if (CheckPositionBeforePlacement(position, fromUser))
         {
-            if (ContractManager.Instance.AddHouse() == false)
+            if (fromUser && ContractManager.Instance.AddHouse() == false)
                 return;
             ContractManager.Instance.updateEditedMap(position, CellType.Structure);
             int randomIndex = GetRandomWeightedIndex(houseWeights);
@@ -48,7 +48,7 @@ public class StructureManager : MonoBehaviour
     {
         if (CheckPositionBeforePlacement(position, fromUser))
         {
-            if (ContractManager.Instance.AddSpecial() == false)
+            if (fromUser && ContractManager.Instance.AddSpecial() == false)
                 return;
             ContractManager.Instance.updateEditedMap(position, CellType.SpecialStructure);
             int randomIndex = GetRandomWeightedIndex(specialWeights);
